requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==3.2
django-cors-headers==3.11.0
django-drf-filepond==0.4.1
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
pycparser==2.21
python-decouple==3.6
pytz==2021.3
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
